+++ packages/ide/src/app/theme.service.ts
@@ -1,23 +1,22 @@
 import { DOCUMENT } from "@angular/common";
 import { ApplicationRef, inject, Service } from "@angular/core";
-import { LocalStorageService } from "ngx-webstorage";
-import { Observable, of, startWith, switchMap } from "rxjs";
-import { SettingsKey } from "../settings";
+import { Observable, of, switchMap } from "rxjs";
+import { settings, ThemePreference } from "../settings";
 import { BrowserThemeService } from "./browser-theme.service";
+import { SettingsService } from "./settings.service";
 
 export type Theme = "light" | "dark";
 
 @Service()
 export class ThemeService {
   private browserTheme = inject(BrowserThemeService);
-  private localStorageSvc = inject(LocalStorageService);
+  private settings = inject(SettingsService);
 
   ref = inject(ApplicationRef);
   document = inject(DOCUMENT);
 
-  theme$ = this.localStorageSvc.observe(SettingsKey.Theme).pipe(
-    startWith(this.localStorageSvc.retrieve(SettingsKey.Theme) || "auto"),
-    switchMap<string, Observable<Theme>>(pref => {
+  theme$ = this.settings.observe(settings.theme).pipe(
+    switchMap<ThemePreference, Observable<Theme>>(pref => {
       if (pref === "light" || pref === "dark") {
         return of(pref);
       }